requirements.txt
# Project Python requirements

# MinIO/S3 connectivity
boto3==1.28.57
botocore==1.31.57

# Spark provider for Airflow
apache-airflow-providers-apache-spark==4.5.0

# Data processing
pandas==2.1.1
pyarrow
